+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.2.0] - 2026-05-29
+
 * [PR-32](https://github.com/OS2web/os2web_datalookup/pull/32)
   Adding Datafordeler address lookup